pyproject.toml
[project]
name = "date-nite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "google-genai>=1.4.0",
    "python-dotenv>=1.2.1",
]
